+++ include/boost/numeric/ublas/tensor/detail/extents_functions.hpp
@@ -156,28 +156,15 @@ namespace boost::numeric::ublas {
 
 /** @brief Returns true if size > 1 and all elements > 0 or size == 1 && e[0] == 1 */
 template <class ExtentsType>
-inline 
+[[nodiscard]] inline 
 constexpr bool valid(ExtentsType const &e) {
   
   static_assert(is_extents_v<ExtentsType>, "boost::numeric::ublas::valid() : invalid type, type should be an extents");
 
-  using size_type = typename ExtentsType::size_type;
-  using value_type = typename ExtentsType::value_type;
+  auto greater_than_zero = [](auto const& a){ return a > 0u; };
 
-  if ( e.size() == size_type(1) ){
-    
-    return e[0] == value_type(1);
-
-  }
-  if ( e.size() > size_type(1) ){
-    
-    return std::all_of(e.begin(), e.end(), [](auto const &a) { return a > value_type(0); });
-  
-  }else{
-
-    return false;
-
-  }
+  if( e.size() == 1u )  return e[0] == 1u;
+  return !e.empty() && std::all_of(e.begin(), e.end(), greater_than_zero );
 }
 
 /**
@@ -188,23 +175,21 @@ constexpr bool valid(ExtentsType const &e) {
  */
 
 template <class T>
-inline
+[[nodiscard]] inline
 std::string to_string(T const &e) {
   
+  using value_type = typename T::value_type;
+
   static_assert(is_extents_v<T> ||is_strides_v<T>, 
     "boost::numeric::ublas::to_string() : invalid type, type should be an extents or a strides");
 
-  if (e.empty()) {
-    return "[]";
-  }
+  if ( e.empty() ) return "[]";
 
   std::stringstream ss;
   
   ss << "[ ";
 
-  for (auto i = typename T::size_type(0); i < e.size() - 1; i++) {
-      ss << e[i] << ", ";
-  }
+  std::copy( e.begin(), e.end() - 1, std::ostream_iterator<value_type>(ss,", ") );
   
   ss << e.back() << " ]";
 
@@ -216,19 +201,14 @@ std::string to_string(T const &e) {
  * @returns true if (1,1,[1,...,1])
  */
 template <class ExtentsType>
-inline 
+[[nodiscard]] inline 
 constexpr bool is_scalar(ExtentsType const &e) {
 
   static_assert(is_extents_v<ExtentsType>, "boost::numeric::ublas::is_scalar() : invalid type, type should be an extents");
+  
+  auto equal_one = [](auto const &a) { return a == 1u; };
 
-  using size_type = typename ExtentsType::size_type;
-  using value_type = typename ExtentsType::value_type;
-
-  if ( e.size() == size_type(0) ){
-    return false;
-  }else{
-    return std::all_of(e.begin(), e.end(), [](auto const &a) { return a == value_type(1); });
-  }
+  return !e.empty() && std::all_of(e.begin(), e.end(), equal_one);
 }
 
 /** @brief Returns true if this has a vector shape
@@ -236,35 +216,19 @@ constexpr bool is_scalar(ExtentsType const &e) {
  * @returns true if (1,n,[1,...,1]) or (n,1,[1,...,1]) with n > 1
  */
 template <class ExtentsType>
-inline 
+[[nodiscard]] inline 
 constexpr bool is_vector(ExtentsType const &e) {
   
   static_assert(is_extents_v<ExtentsType>, "boost::numeric::ublas::is_vector() : invalid type, type should be an extents");
 
-  using size_type = typename ExtentsType::size_type;
-  using value_type = typename ExtentsType::value_type; 
+  auto greater_one = [](auto const &a) { return a > 1u; };
+  auto equal_one = [](auto const &a) { return a == 1u; };
 
-  if (e.size() == size_type(0)) {
-
-    return false;
-
-  } else if (e.size() == size_type(1)) {
-
-    return e.at(0) > value_type(1);
-    
-  }else{
-
-    auto greater_one = [](auto const &a) {
-      return a > value_type(1);
-    };
-    auto equal_one = [](auto const &a) { 
-      return a == value_type(1); 
-    };
-
-    return std::any_of(e.begin(), e.begin() + 2, greater_one) &&
-            std::any_of(e.begin(), e.begin() + 2, equal_one) &&
-            std::all_of(e.begin() + 2, e.end(), equal_one);
-  }
+  if      (e.size() == 0u) return false;
+  else if (e.size() == 1u) return e[0] > 1u;
+  else  return  std::any_of(e.begin(), e.begin() + 2, greater_one) &&
+                std::any_of(e.begin(), e.begin() + 2, equal_one) &&
+                std::all_of(e.begin() + 2, e.end(), equal_one);
 
 }
 
@@ -273,30 +237,16 @@ constexpr bool is_vector(ExtentsType const &e) {
  * @returns true if (m,n,[1,...,1]) with m > 1 and n > 1
  */
 template <class ExtentsType>
-inline 
+[[nodiscard]] inline 
 constexpr bool is_matrix(ExtentsType const &e) {
   
   static_assert(is_extents_v<ExtentsType>, "boost::numeric::ublas::is_matrix() : invalid type, type should be an extents");
 
-  using size_type = typename ExtentsType::size_type;
-  using value_type = typename ExtentsType::value_type;
+  auto greater_one = [](auto const &a) { return a > 1u; };
+  auto equal_one = [](auto const &a) { return a == 1u; };
 
-  if (e.size() < size_type(2)) {
-    return false;
-  }else{
-
-    auto greater_one = [](auto const &a) {
-      return a > value_type(1);
-    };
-
-    auto equal_one = [](auto const &a) { 
-      return a == value_type(1); 
-    };
-
-    return std::all_of(e.begin(), e.begin() + 2, greater_one) &&
-            std::all_of(e.begin() + 2, e.end(), equal_one);
-  }
-
+  return  ( e.size() >= 2u ) && std::all_of(e.begin(), e.begin() + 2, greater_one) &&
+          std::all_of(e.begin() + 2, e.end(), equal_one);
 }
 
 /** @brief Returns true if this is has a tensor shape
@@ -304,28 +254,14 @@ constexpr bool is_matrix(ExtentsType const &e) {
  * @returns true if !empty() && !is_scalar() && !is_vector() && !is_matrix()
  */
 template <class ExtentsType>
-inline 
+[[nodiscard]] inline 
 constexpr bool is_tensor(ExtentsType const &e) {
 
   static_assert(is_extents_v<ExtentsType>, "boost::numeric::ublas::is_tensor() : invalid type, type should be an extents");
   
-  using size_type = typename ExtentsType::size_type;
-  using value_type = typename ExtentsType::value_type;
+  auto greater_one = [](auto const &a) { return a > 1u;};
   
-  if (e.size() < size_type(3)) {
-    
-    return false;
-
-  }else{
-
-    auto greater_one = [](auto const &a) {
-      return a > value_type(1);
-    };
-
-    return std::any_of(e.begin() + 2, e.end(), greater_one);
-
-  }
-
+  return  ( e.size() >= 3u ) && std::any_of(e.begin() + 2, e.end(), greater_one);
 }
 
 /** @brief Eliminates singleton dimensions when size > 2
@@ -341,7 +277,7 @@ constexpr bool is_tensor(ExtentsType const &e) {
  * @returns basic_extents<int_type> with squeezed extents
  */
 template <class ExtentsType>
-inline
+[[nodiscard]] inline
 auto squeeze(ExtentsType const &e) {
   
   static_assert(is_extents_v<ExtentsType>, "boost::numeric::ublas::squeeze() : invalid type, type should be an extents");
@@ -351,23 +287,13 @@ auto squeeze(ExtentsType const &e) {
 
 /** @brief Returns the product of extents */
 template <class ExtentsType>
-inline
+[[nodiscard]] inline
 constexpr auto product(ExtentsType const &e) {
 
   static_assert(is_extents_v<ExtentsType>, "boost::numeric::ublas::product() : invalid type, type should be an extents");
   
-  using value_type = typename ExtentsType::value_type;
-  
-  if (e.empty()) {
-
-    return value_type(0);
-    
-  }else{
-
-    return value_type( std::accumulate(e.begin(), e.end(), value_type(1), std::multiplies<>()) );
-
-  }
-
+  if ( e.empty() ) return 0u;
+  else return std::accumulate(e.begin(), e.end(), 1u, std::multiplies<>()) ;
 }
 
 
@@ -378,6 +304,10 @@ template <class LExtents, class RExtents,
 >
 [[nodiscard]] inline
 constexpr bool operator==(LExtents const& lhs, RExtents const& rhs) noexcept{
+  
+  static_assert( std::is_same_v<typename LExtents::value_type, typename RExtents::value_type>, 
+    "boost::numeric::ublas::operator==(LExtents, RExtents) : LHS value type should be same as RHS value type");
+
   return ( lhs.size() == rhs.size() ) && std::equal(lhs.begin(), lhs.end(), rhs.begin());
 }
 
@@ -388,6 +318,10 @@ template <class LExtents, class RExtents,
 >
 [[nodiscard]] inline
 constexpr bool operator!=(LExtents const& lhs, RExtents const& rhs) noexcept{
+  
+  static_assert( std::is_same_v<typename LExtents::value_type, typename RExtents::value_type>, 
+    "boost::numeric::ublas::operator!=(LExtents, RExtents) : LHS value type should be same as RHS value type");
+
   return !( lhs == rhs );
 }
