+++ include/boost/numeric/ublas/tensor/detail/type_traits_fixed_rank_extents.hpp
@@ -13,7 +13,7 @@
 #ifndef BOOST_UBLAS_TENSOR_TYPE_TRAITS_FIXED_RANK_EXTENTS_HPP
 #define BOOST_UBLAS_TENSOR_TYPE_TRAITS_FIXED_RANK_EXTENTS_HPP
 
-#include <boost/numeric/ublas/tensor/detail/type_traits_extents.hpp>
+#include <boost/numeric/ublas/tensor/detail/basic_type_traits.hpp>
 
 namespace boost::numeric::ublas{
